refactor: split auth guards and protected paths
